Create APIClass
In this question, we will be creating a class API using ES6 class syntax.

The class will have 2 public fields, url, and method. The constructor method will take the URL as a parameter but the method argument is default to 'GET'.

Also, the class object will have one private field (use #syntax) secure which will be set on initialization to true if the URL starts with https else will be set to false.

The class will have two public methods, isSecure which returns the object's #secure properties' value, and a updateURL method which updates the URL and also updates the #secure field accordingly.